KingStoreV/requirements.txt
aiogram==3.1.0
aiosqlite>=0.17.0
python-dotenv>=1.0.0
pydantic>=1.10
pytest>=7.0
requests>=2.28